Remove the `collapsed` and `scrolled` fields from cell metadata.
This fields control to display the output of a code cell and can be
changed from the notebook editor.

They can be annoying in the diffs. Besides, since the two properties
are reset when the cell is executed and the `ipynb_output_filter`
script is removing already the output of cells, there is no point
in keeping them.